Add files via upload
The __MACOSX folder is a hidden directory that contains extra information about files and folders in zipped folders. It's used by macOS, the operating system that runs on Apple Mac computers
macOS automatically adds the __MACOSX folder when creating a zip. 
It's only designed to be used with macOS and has no purpose on other operating systems. 
here  __MACOSX folder is being downloaded via zip file in windows causing trouble error when installing react packages.
so here is a zip file not having _MACOSX folder